Renamed "dlg_add_set.*" to "dlg_add_set_result.*" and updated references.
