+++ cockatrice/src/dlg_add_set_result.h
